test: Remove unnecessary registration test
Registration is already tested in the main integration test
